GVT-1733: Refactor vertical geometry listing generation
