+++ infra/src/main/kotlin/fi/fta/geoviite/infra/geometry/GeometryService.kt
@@ -2,8 +2,10 @@ package fi.fta.geoviite.infra.geometry
 
 import fi.fta.geoviite.infra.common.*
 import fi.fta.geoviite.infra.common.PublishType.OFFICIAL
+import fi.fta.geoviite.infra.geocoding.GeocodingContext
 import fi.fta.geoviite.infra.geocoding.GeocodingService
 import fi.fta.geoviite.infra.geography.CoordinateTransformationService
+import fi.fta.geoviite.infra.geography.Transformation
 import fi.fta.geoviite.infra.geometry.PlanSource.PAIKANNUSPALVELU
 import fi.fta.geoviite.infra.inframodel.InfraModelFile
 import fi.fta.geoviite.infra.logging.serviceCall
@@ -222,7 +224,7 @@ class GeometryService @Autowired constructor(
         return FileName("$ELEMENT_LISTING ${track.name}") to csvFileContent.toByteArray()
     }
 
-    fun getGeometryProfile(
+    fun getVerticalGeometryListing(
         planId: IntId<GeometryPlan>
     ): List<VerticalGeometryListing> {
         val planHeader = getPlanHeader(planId)
@@ -230,79 +232,17 @@ class GeometryService @Autowired constructor(
         val geocodingContext = geocodingService.getGeocodingContext(OFFICIAL, planHeader.trackNumberId)
         val coordinateTransform = planHeader.units.coordinateSystemSrid?.let(coordinateTransformationService::getLayoutTransformation)
 
-        return alignments.filter { it.profile != null }.map { alignment ->
-            val (curvedSegments, linearSegments) =
-                alignment.profile?.segments?.partition { it is CurvedProfileSegment }
-                    ?.let { partitioned ->
-                        partitioned.first.map { it as CurvedProfileSegment } to
-                                partitioned.second.map { it as LinearProfileSegment } }
-                    ?: (emptyList<CurvedProfileSegment>() to emptyList())
-            curvedSegments.map { segment ->
-                    toVerticalGeometryListing(
-                        segment,
-                        alignment,
-                        null,
-                        coordinateTransform,
-                        planHeader.id,
-                        planHeader.source,
-                        planHeader.fileName,
-                        geocodingContext,
-                        curvedSegments,
-                        linearSegments
-                    )
-            }
-        }.flatten()
+        return toVerticalGeometryListing(alignments, coordinateTransformationService::getLayoutTransformation, planHeader, geocodingContext)
     }
 
-    data class GeometryProfileCalculationContext(
-        val geometryAlignment: GeometryAlignment,
-        val planHeader: GeometryPlanHeader,
-        val curvedProfileSegments: List<CurvedProfileSegment>,
-        val linearProfileSegments: List<LinearProfileSegment>,
-    )
-
-    fun getGeometryProfile(
+    fun getVerticalGeometryListing(
         locationTrackId: IntId<LocationTrack>,
         startAddress: TrackMeter?,
         endAddress: TrackMeter?,
     ): List<VerticalGeometryListing> {
-        val (track, alignment) = locationTrackService.getWithAlignmentOrThrow(OFFICIAL, locationTrackId)
-        val geocodingContext = geocodingService.getGeocodingContext(OFFICIAL, track.trackNumberId)
-        val linkedElementIds = collectLinkedElements(
-            alignment.segments,
-            geocodingContext,
-            startAddress,
-            endAddress
-        ).mapNotNull { it.second }
-        val headersAndAlignments = linkedElementIds
-            .map(::getAlignmentId)
-            .distinct()
-            .associateWith(::getHeaderAndAlignment)
-
-        val curvedSegmentsAndGeometryListingContexts = linkedElementIds
-            .map { elementId ->
-                getCurvedProfileSegmentsAndContextsOverlappingElement(headersAndAlignments, elementId)
-            }
-            .flatten()
-            .distinctBy { it.first }
-
-        return curvedSegmentsAndGeometryListingContexts.map { (segment, context) ->
-            toVerticalGeometryListing(
-                segment,
-                context.geometryAlignment,
-                track.name,
-                context.planHeader.units.coordinateSystemSrid
-                    ?.let { coordinateTransformationService
-                        .getLayoutTransformation(context.planHeader.units.coordinateSystemSrid)
-                          },
-                context.planHeader.id,
-                context.planHeader.source,
-                context.planHeader.fileName,
-                geocodingContext,
-                context.curvedProfileSegments,
-                context.linearProfileSegments
-            )
-        }
+        val (track, alignment) = locationTrackService.getWithAlignmentOrThrow(PublishType.OFFICIAL, locationTrackId)
+        val geocodingContext = geocodingService.getGeocodingContext(PublishType.OFFICIAL, track.trackNumberId)
+        return toVerticalGeometryListing(track, alignment, startAddress, endAddress, geocodingContext, coordinateTransformationService::getLayoutTransformation, ::getHeaderAndAlignment)
     }
 
     private fun getHeaderAndAlignment(id: IntId<GeometryAlignment>): Pair<GeometryPlanHeader, GeometryAlignment> {
