+++ infra/src/main/kotlin/fi/fta/geoviite/infra/geometry/VerticalGeometryListing.kt
@@ -4,8 +4,17 @@ import fi.fta.geoviite.infra.common.*
 import fi.fta.geoviite.infra.geocoding.GeocodingContext
 import fi.fta.geoviite.infra.geography.Transformation
 import fi.fta.geoviite.infra.math.*
+import fi.fta.geoviite.infra.tracklayout.LayoutAlignment
+import fi.fta.geoviite.infra.tracklayout.LocationTrack
 import fi.fta.geoviite.infra.util.FileName
 import kotlin.math.tan
+
+data class GeometryProfileCalculationContext(
+    val geometryAlignment: GeometryAlignment,
+    val planHeader: GeometryPlanHeader,
+    val curvedProfileSegments: List<CurvedProfileSegment>,
+    val linearProfileSegments: List<LinearProfileSegment>,
+)
 
 data class CurvedSectionEndpoint(
     val address: TrackMeter?,
@@ -43,6 +52,78 @@ data class VerticalGeometryListing(
     val linearSectionForward: LinearSection,
     val linearSectionBackward: LinearSection,
 )
+
+fun toVerticalGeometryListing(
+    planAlignments: List<GeometryAlignment>,
+    getTransformation: (srid: Srid) -> Transformation,
+    planHeader: GeometryPlanHeader,
+    geocodingContext: GeocodingContext?
+): List<VerticalGeometryListing> {
+    return planAlignments.filter { it.profile != null }.map { alignment ->
+        val (curvedSegments, linearSegments) =
+            alignment.profile?.segments
+                ?.let(::separateCurvedAndLinearProfileSegments)
+                ?: (emptyList<CurvedProfileSegment>() to emptyList())
+        curvedSegments.map { segment ->
+            toVerticalGeometryListing(
+                segment,
+                alignment,
+                null,
+                planHeader.units.coordinateSystemSrid?.let(getTransformation),
+                planHeader.id,
+                planHeader.source,
+                planHeader.fileName,
+                geocodingContext,
+                curvedSegments,
+                linearSegments
+            )
+        }
+    }.flatten()
+}
+
+fun toVerticalGeometryListing(
+    track: LocationTrack,
+    layoutAlignment: LayoutAlignment,
+    startAddress: TrackMeter?,
+    endAddress: TrackMeter?,
+    geocodingContext: GeocodingContext?,
+    getTransformation: (srid: Srid) -> Transformation,
+    getPlanHeaderAndAlignment: (id: IntId<GeometryAlignment>) -> Pair<GeometryPlanHeader, GeometryAlignment>,
+): List<VerticalGeometryListing> {
+    val linkedElementIds = collectLinkedElements(
+        layoutAlignment.segments,
+        geocodingContext,
+        startAddress,
+        endAddress
+    ).mapNotNull { it.second }
+    val headersAndAlignments = linkedElementIds
+        .map(::getAlignmentId)
+        .distinct()
+        .associateWith(getPlanHeaderAndAlignment)
+
+    val curvedSegmentsAndGeometryListingContexts = linkedElementIds
+        .map { elementId ->
+            getCurvedProfileSegmentsAndContextsOverlappingElement(headersAndAlignments, elementId)
+        }
+        .flatten()
+        .distinctBy { it.first }
+
+    return curvedSegmentsAndGeometryListingContexts.map { (segment, context) ->
+        toVerticalGeometryListing(
+            segment,
+            context.geometryAlignment,
+            track.name,
+            context.planHeader.units.coordinateSystemSrid
+                ?.let(getTransformation),
+            context.planHeader.id,
+            context.planHeader.source,
+            context.planHeader.fileName,
+            geocodingContext,
+            context.curvedProfileSegments,
+            context.linearProfileSegments
+        )
+    }
+}
 
 fun toVerticalGeometryListing(
     segment: CurvedProfileSegment,
@@ -157,7 +238,7 @@ fun angleFractionBetweenPoints(point1: IPoint, point2: IPoint) =
 fun getCurvedProfileSegmentsAndContextsOverlappingElement(
     headersAndAlignments: Map<IntId<GeometryAlignment>, Pair<GeometryPlanHeader, GeometryAlignment>>,
     elementId: IndexedId<GeometryElement>
-): List<Pair<CurvedProfileSegment, GeometryService.GeometryProfileCalculationContext>> {
+): List<Pair<CurvedProfileSegment, GeometryProfileCalculationContext>> {
     val (planHeader, geometryAlignment) = headersAndAlignments.getValue(getAlignmentId(elementId))
     val (curvedSegments, linearSegments) =
         geometryAlignment.profile?.segments
@@ -171,7 +252,7 @@ fun getCurvedProfileSegmentsAndContextsOverlappingElement(
         }
 
     return segmentsOverlappingElement.map { curve ->
-        curve to GeometryService.GeometryProfileCalculationContext(
+        curve to GeometryProfileCalculationContext(
             geometryAlignment,
             planHeader,
             curvedSegments,
